+++ Consequences/Receptors/DamageResult.cs
@@ -1,10 +1,10 @@
 namespace Consequences.Receptors;
 
 public readonly record struct DamageResult(
-    double Structure,
-    double Content)
+    float Structure,
+    float Content)
 {
     public const int ComponentCount = 2;
 
-    public double Total => Structure + Content;
+    public float Total => Structure + Content;
 }
